The code I wrote while testing out tokio.
I came to the solution in test10/11 but it didn't fix my issue after all.
My issue was that I wanted to call some sync code that blocks_on async gRPC calls. And the sync call was going to be called in async context by another crate (that's why I couldn't use spawn_blocking; the crate calling my sync code should be the one calling spawn_blocking on it, but that wasn't possible either, as other users don't have such constrains).

Why test10/11 didn't fix my issue was because I made a mistake while placing each thing in a runtime in the original code piece. \`tokio::sleep`\ is the call I used to simulate an async gRPC call, but a gRPC call is different than a sleep, because it calls on another tokio thread - the gRPC server - and waits for its response, while the sleep does non of that.

The mistake I did in the original code piece was placing both the gRPC server and and the code that blocks_on client calls in the same tokio thread, seperating them was the solution: put the server in the global runtime and block_on the client calls using the global runtime handle, but in an artificial/extra runtime.